Major refactor.
* Each CLI subcommand is now in its own module.  Common code has been
  moved to other modules, especially the ones under `internal/`.

* Breaking changes to CLI!  But now each CLI command has its own flags
  and its own help command, so we don't need a giant confusing jumble of
  flags for every command.

* The "apply" subcommand now uses a config file (JSON or YAML,
  autodetected by file suffix) to list the Mastodon instances to apply
  to.

* The "apply" subcommand now supports the REST API, but it requires a
  bearer token.

* The "register" subcommand now exists, for obtaining a bearer token.

This PR will need to be paired with some additional changes in the APT
repo.  Hopefully this will be the last major change needed to the APT
repo, as the new `dist/debian.manifest.json` file will be used to script
the copying of files into the Debian package, rather than relying on
keeping the APT Python script up-to-date.